Clean up

+++ src/main/kotlin/net/barribob/boss/utils/VanillaCopies.kt
@@ -112,8 +112,8 @@ object VanillaCopies {
             val i: Int = packet.id
             entity15.updateTrackedPosition(d, e, f)
             entity15.refreshPositionAfterTeleport(d, e, f)
-            entity15.pitch = (packet.pitch * 360).toFloat() / 256.0f
-            entity15.yaw = (packet.yaw * 360).toFloat() / 256.0f
+            entity15.pitch = (packet.pitch * 360) / 256.0f
+            entity15.yaw = (packet.yaw * 360) / 256.0f
             entity15.id = i
             entity15.uuid = packet.uuid
             world.addEntity(i, entity15 as Entity?)
